+++ ouzel/core/android/main.cpp
@@ -52,7 +52,8 @@ extern "C" JNIEXPORT void JNICALL Java_org_ouzel_OuzelLibJNIWrapper_onSurfaceDes
 extern "C" JNIEXPORT void JNICALL Java_org_ouzel_OuzelLibJNIWrapper_onSurfaceChanged(JNIEnv*, jclass, jobject, jint width, jint height)
 {
     ouzel::NativeWindowAndroid* windowAndroid = static_cast<ouzel::NativeWindowAndroid*>(engine->getWindow()->getNativeWindow());
-    windowAndroid->handleResize(ouzel::Size2<float>(static_cast<float>(width), static_cast<float>(height)));
+    windowAndroid->handleResize(ouzel::Size2<uint32_t>(static_cast<uint32_t>(width),
+                                                       static_cast<uint32_t>(height)));
 }
 
 extern "C" JNIEXPORT void JNICALL Java_org_ouzel_OuzelLibJNIWrapper_onStart(JNIEnv*, jclass)
