+++ ouzel/core/tvos/NativeWindowTVOS.mm
@@ -32,8 +32,8 @@
 {
     [super viewWillTransitionToSize:size withTransitionCoordinator:coordinator];
 
-    window->handleResize(ouzel::Size2<float>(static_cast<float>(size.width),
-                                      static_cast<float>(size.height)));
+    window->handleResize(ouzel::Size2<uint32_t>(static_cast<uint32_t>(size.width),
+                                                static_cast<uint32_t>(size.height)));
 }
 
 @end
@@ -45,7 +45,7 @@ namespace ouzel
                                        graphics::Driver graphicsDriver,
                                        bool newHighDpi):
         NativeWindow(initCallback,
-                     Size2<float>(),
+                     Size2<uint32_t>(),
                      true,
                      true,
                      true,
@@ -61,8 +61,8 @@ namespace ouzel
 
         CGRect windowFrame = [window bounds];
 
-        size = Size2<float>(static_cast<float>(windowFrame.size.width),
-                     static_cast<float>(windowFrame.size.height));
+        size = Size2<uint32_t>(static_cast<uint32_t>(windowFrame.size.width),
+                               static_cast<uint32_t>(windowFrame.size.height));
 
         switch (graphicsDriver)
         {
@@ -96,7 +96,7 @@ namespace ouzel
         if (highDpi)
         {
             contentScale = static_cast<float>(screen.scale);
-            resolution = size * contentScale;
+            resolution = size * static_cast<uint32_t>(contentScale);
         }
         else
             resolution = size;
@@ -127,7 +127,7 @@ namespace ouzel
         }
     }
 
-    void NativeWindowTVOS::handleResize(const Size2<float>& newSize)
+    void NativeWindowTVOS::handleResize(const Size2<uint32_t>& newSize)
     {
         size = newSize;
         resolution = size * contentScale;
