+++ ouzel/graphics/opengl/OGLRenderDevice.cpp
@@ -425,7 +425,7 @@ namespace ouzel
         }
 
         void OGLRenderDevice::init(Window* newWindow,
-                                   const Size2<float>& newSize,
+                                   const Size2<uint32_t>& newSize,
                                    uint32_t newSampleCount,
                                    Texture::Filter newTextureFilter,
                                    uint32_t newMaxAnisotropy,
@@ -865,7 +865,7 @@ namespace ouzel
             clearDepth = newClearDepth;
         }
 
-        void OGLRenderDevice::setSize(const Size2<float>& newSize)
+        void OGLRenderDevice::setSize(const Size2<uint32_t>& newSize)
         {
             RenderDevice::setSize(newSize);
