+++ ouzel/graphics/opengl/windows/OGLRenderDeviceWin.hpp
@@ -28,7 +28,7 @@ namespace ouzel
 
         private:
             void init(Window* newWindow,
-                      const Size2<float>& newSize,
+                      const Size2<uint32_t>& newSize,
                       uint32_t newSampleCount,
                       Texture::Filter newTextureFilter,
                       uint32_t newMaxAnisotropy,
